Add Firefox native messaging manifest installer
Renamed install-for-chrome-ext to install-ext-host, added Firefox
support and added a new deprecated command alias for the old name.
