+++ site/content/docs/browser-extension.md
@@ -15,10 +15,6 @@ The Sail browser extension allows you to open GitHub or GitLab projects with a s
 ## Install
 
 1. [Install Sail if you haven't already](/docs/installation)
-1. Run `sail install-for-chrome-ext` to install the chrome extension manifest.json
+1. Run `sail install-ext-host` to install the extension manifest.json
 1. [Install the extension from the Chrome Marketplace](https://chrome.google.com/webstore/detail/sail/deeepphleikpinikcbjplcgojfhkcmna)
 1. Get Sailing!
-
-
-
-
